requirements.txt
Flask
openai
python-dotenv
psycopg2-binary
gunicorn
PyPDF2
reportlab
weasyprint
stripe
flask_sqlalchemy
flask_login
flask_migrate
authlib
requests
PyMuPDF
python-docx